refactor: Move maps utils KTX into a different directory
BREAKING CHANGE: Moved KTX for utils from com.google.maps.android.ktx to
com.google.maps.android.ktx.utils to avoid package name conflicts with
the new maps-ktx module.
