+++ maps-utils-ktx/src/main/java/com/google/maps/android/ktx/utils/core/PolygonOptions.kt
@@ -15,7 +15,7 @@
  *
  */
 
-package com.google.maps.android.ktx.core
+package com.google.maps.android.ktx.utils.core
 
 import com.google.android.gms.maps.model.PolygonOptions
