+++ maps-utils-ktx/src/main/java/com/google/maps/android/ktx/utils/kml/Kml.kt
@@ -16,7 +16,7 @@
  */
 
 @file:Suppress("NOTHING_TO_INLINE")
-package com.google.maps.android.ktx.kml
+package com.google.maps.android.ktx.utils.kml
 
 import android.content.Context
 import androidx.annotation.RawRes
