+++ maps-utils-ktx/src/test/java/com/google/maps/android/ktx/utils/LatLngTest.kt
@@ -15,7 +15,7 @@
  *
  */
 
-package com.google.maps.android.ktx
+package com.google.maps.android.ktx.utils
 
 import com.google.android.gms.maps.model.LatLng
 import org.junit.Assert.*
