+++ maps-utils-ktx/src/test/java/com/google/maps/android/ktx/utils/PolylineTest.kt
@@ -15,10 +15,12 @@
  *
  */
 
-package com.google.maps.android.ktx
+package com.google.maps.android.ktx.utils
 
 import com.google.android.gms.maps.model.LatLng
 import com.google.android.gms.maps.model.Polyline
+import com.google.maps.android.ktx.utils.contains
+import com.google.maps.android.ktx.utils.sphericalPathLength
 import com.nhaarman.mockitokotlin2.doReturn
 import com.nhaarman.mockitokotlin2.mock
 import org.junit.Assert.*
