+++ maps-utils-ktx/src/test/java/com/google/maps/android/ktx/utils/core/PolygonOptionsTest.kt
@@ -15,7 +15,7 @@
  *
  */
 
-package com.google.maps.android.ktx.core
+package com.google.maps.android.ktx.utils.core
 
 import android.graphics.Color
 import com.google.android.gms.maps.model.LatLng
@@ -25,11 +25,12 @@ import org.junit.Test
 class PolygonOptionsTest {
     @Test
     fun testBuilder() {
-        val polygonOptions = buildPolygonOptions {
-            strokeWidth(1.0f)
-            strokeColor(Color.BLACK)
-            add(LatLng(0.0, 0.0))
-        }
+        val polygonOptions =
+            buildPolygonOptions {
+                strokeWidth(1.0f)
+                strokeColor(Color.BLACK)
+                add(LatLng(0.0, 0.0))
+            }
         assertEquals(1.0f, polygonOptions.strokeWidth)
         assertEquals(Color.BLACK, polygonOptions.strokeColor)
         assertEquals(listOf(LatLng(0.0, 0.0)), polygonOptions.points)
